contrail-container-builder fails to initialize vhost0 with bond in tripleo
Closes-Bug: 1777996

+++ containers/vrouter/base/agent-functions.sh
@@ -67,6 +67,7 @@ function get_bonding_parameters() {
         local mode="$(cat ${bond_dir}/mode | awk '{print $2}')"
         local policy="$(cat ${bond_dir}/xmit_hash_policy | awk '{print $1}')"
         policy=$(convert_bond_policy $policy)
+        mode=$(convert_bond_mode $mode)
 
         local slaves="$(cat ${bond_dir}/slaves | tr ' ' '\n' | sort | tr '\n' ',')"
         slaves=${slaves%,}
@@ -346,6 +347,7 @@ function read_and_save_dpdk_params() {
         if [ -n "$BOND_MODE" ] ; then
             mode=$BOND_MODE
             policy=$(convert_bond_policy $BOND_POLICY)
+            mode=$(convert_bond_mode ${mode})
             slaves=''
             declare _slave
             for _slave in ${BIND_INT//,/ } ; do
@@ -461,7 +463,7 @@ function init_vhost0() {
     fi
 
     local ret=0
-    if [[ -e /etc/sysconfig/network-scripts/ifcfg-${phys_int} ]]; then
+    if [[ -e /etc/sysconfig/network-scripts/ifcfg-${phys_int} || -e /etc/sysconfig/network-scripts/ifcfg-vhost0 ]]; then
         echo "INFO: creating ifcfg-vhost0 and initialize it via ifup"
         if ! is_dpdk ; then
             ifdown ${phys_int}
@@ -475,7 +477,7 @@ function init_vhost0() {
             /bin/cp -f route-${phys_int} route-vhost0
             mv route-${phys_int} contrail.org.route-${phys_int}
         fi
-        if [ ! -f "contrail.org.ifcfg-${phys_int}" ] ; then
+        if [[ ! -f "contrail.org.ifcfg-${phys_int}" && -f "ifcfg-${phys_int}" ]] ; then
             /bin/cp -f ifcfg-${phys_int} contrail.org.ifcfg-${phys_int}
             sed -r "/(DEVICE|TYPE|ONBOOT|MACADDR|HWADDR|BONDING|SLAVE|VLAN|MTU)/! s/^[^#].*/#commented_by_contrail& /" ifcfg-${phys_int} > ifcfg-${phys_int}.tmp
             echo 'NM_CONTROLLED=no' >> ifcfg-${phys_int}.tmp
@@ -500,7 +502,7 @@ function init_vhost0() {
             ip route del $line || { echo "ERROR: route $line was not removed for iface ${phys_int}." && ret=1; }
         done < <(ip route sh | grep ${phys_int})
     else
-        echo "INFO: there is no ifcfg-$phys_int, so initialize vhost0 manually"
+        echo "INFO: there is no ifcfg-$phys_int and ifcfg-vhost0, so initialize vhost0 manually"
         # TODO: switch off dhcp on phys_int
         echo "INFO: Changing physical interface to vhost in ip table"
         echo "$addrs" | while IFS= read -r line ; do
